+++ lib/msf/ui/console/driver.rb
@@ -30,6 +30,7 @@ class Driver < Msf::Ui::Driver
   # Console Command Dispatchers to be loaded after the Core dispatcher.
   #
   CommandDispatchers = [
+    CommandDispatcher::Modules,
     CommandDispatcher::Jobs,
     CommandDispatcher::Resource
   ]
